Use a single key to store deltas
To improve the distribution of the stored deltas, we removed the deep
key value maps and use a custom key to encode the `context` and
`timestamp` of a stored delta. Thus, improving the amount of data
that needs to be loaded for each operation.

Added additional end-to-end tests and benchmarks to validate the
behavior.

+++ src/main.rs
@@ -330,7 +330,7 @@ fn parse_date(string: &str) -> Result<Date, time::error::Parse> {
 #[derive(Deserialize)]
 #[serde(rename_all = "camelCase")]
 struct DeltaScopeInput {
-    context: Option<u64>,
+    context: Option<u32>,
     date: String,
 }
